requirements-dev.txt
-r requirements.txt

# Development dependencies
black>=22.3.0
flake8>=4.0.1
isort>=5.10.1
pytest-cov>=3.0.0 